Fix deployement & test
