--- tests/updates/nominal_case/2020-02/1-optimize-weight-assets.md
@@ -1,19 +1,0 @@
----
-pim_announcement_img: ./img/Optimize_jpeg_transformation.png
-pim_announcement_alt_img: "this is alt img"
-pim_announcement_audience:
-- EE
----
-
-# Optimize the weight of your assets
-::: meta-data type="Improvement" features="Asset Manager" available="late February" link-to-doc="../articles/assets-transformation.html#optimize-jpeg"
-:::
-
-Before applying transformations on your assets, we first convert your files into png to avoid any image quality loss. This conversion can increase the size of your media files. If you need to reduce the size of your images, you can now use a new operation `optimize_jpeg`. This transformation operation will compress your file to a jpeg. Thanks to this new feature, you can easily optimize the weight of your new images and thus save asset storage space.
-
-![Optimize jpeg transformation](../img/Optimize_jpeg_transformation.png)
-
-::: more    
-[What is a transformation?](../articles/assets-transformation.html)  
-[What are the available asset transformation operations?](/articles/assets-transformation.html#transformation-operations)
-:::
