--- tests/updates/nominal_case/2020-02/2-new-warning-message-username-connection.md
@@ -1,22 +1,0 @@
----
-pim_announcement_audience:
-- CE
-- EE
----
-
-# New warning message on connection usernames
-::: meta-data type="Improvement" features="Connections" available="late February" link-to-doc="../articles/manage-your-connections.html#why-should-you-use-the-connection-username"
-:::
-
-Whenever one of your connectors is not using the dedicated connection username, we display a warning message in the main Connections screen as well as in the Connection form.     
-This warning message mentions which usernames were wrongly used for this connection and when they were used for the last time.    
-Thanks to this improvement, we ensure the API connection security and accesses.
-
-Want to discover why we need to guarantee this? Read our help center section about [why it is crucial to use the username of your connection](../articles/manage-your-connections.html#why-should-you-use-the-connection-username).
-
-![Connection user warning](../img/connection_user_warning_message.png)
-
-::: more
-[What is a connection?](../articles/what-is-a-connection.html)  
-[How to manage a connection?](../articles/manage-your-connections.html)
-:::
