--- tests/updates/nominal_case/2020-03/1-data-quality-new-improvements.md
@@ -1,42 +1,0 @@
----
-pim_announcement_img: ./img/dqi_variant_products.png
-pim_announcement_alt_img: "this is DQI alt img"
-pim_announcement_audience:
-- CE
-- EE
----
-
-# Data quality improvements
-
-This is a data quality improvement description.
-
-## Data Quality Insights now available for variant products
-::: meta-data type="Improvement" features="Data Quality Insights, Data Quality" available="early March" link-to-doc="../articles/improve-data-quality.html#improve-data-quality-on-variant-products"
-:::
-
-To help you measure and improve your product data quality, **our Data Quality Insights recommendation tab and grades are now available for your variant products**.
-
-Get specific enrichment and consistency grading scores for all your product variants and quickly identify the list of attributes you need to improve for a specific variant product whether these attributes are at the product model or the variant level.
-
-![Data Quality Insights on variant products](../img/dqi_variant_products.png)
-
-## Easy and direct access to attributes to be improved
-::: meta-data type="Improvement" features="Data Quality Insights, Data Quality" available="early March" link-to-doc="../articles/improve-data-quality.html#improve-data-quality"
-:::
-
-If you have any improvement recommendations under the Data Quality Insights tab, simply **click on your attribute to directly access it and improve it**!
-
-![Ease data quality improvements in the PEF](../img/ease_data_quality_improvements_in_pef.png)
-
-## Spell-check improvements
-::: meta-data type="Improvement" features="Data Quality Insights, Spell-check, Data Quality" available="early March" link-to-doc="../articles/improve-data-quality.html#improve-data-quality"
-:::
-
-To help reach a higher data quality content yet, **spell-check and suggestions are now available on text area attributes with rich text editor enabled** (WYSIWYG).
-
-The spell-check feature now **supports four new languages**: Italian 🇮🇹, Swedish 🇸🇪, Danish 🇩🇰 and Dutch 🇳🇱. This means that you can now have spelling mistakes automatically detected for no more than 8 locales. Mistakes are no longer allowed! :wink:
-
-::: more
-[What is data quality?](../articles/understand-data-quality.html)  
-[How to improve your data quality?](../articles/improve-data-quality.html#improve-data-quality)
-:::
