+++ tests/updates/nominal_case/asset-manager/1-greater-flexibility.md
@@ -1,0 +1,10 @@
+# With greater flexibility come greater possibilities
+
+Assets are now much more flexible and can be managed and structured in the same ways as products and reference entities. They now belong to asset families and have attributes, and just like products, assets belonging
+to an asset family inherit its attributes.
+
+![The new asset form](../img/Asset_form.png)
+
+You can define as many asset families as you need, and several types of attributes are available.  
+You can structure your families with several attribute types: text, single and multiple options, number, media file, and media link. The media file attribute allows you to upload any file format, including images in PNG, JPG, TIF,
+and PSD format, as well as PDFs, Word documents, and more. The media link attribute lets you add and preview media files hosted externally, such as an image from a website or a YouTube video.
