+++ tests/updates/nominal_case/asset-manager/2-3rd-party-media.md
@@ -1,0 +1,5 @@
+# Easily retrieve 3rd party media
+
+New API endpoints enable the connection between Akeneo and your existing media stores. These endpoints allow you to build your own middleware to retrieve media from your DAM, media storage, or any other external source. Assets stored externally will also now be visible along with assets uploaded to the PIM asset library.
+
+In order to offer the flexibility to connect to any 3rd party media storage, we provide extensive documentation explaining [how to connect your DAM to Akeneo PIM](https://api.akeneo.com/guides/dam-connection/introduction.html). This online resource contains a step-by-step guide to develop the middleware, and a middleware skeleton to help you easily and quickly connect Akeneo to your media storage.
