+++ tests/updates/nominal_case/asset-manager/3-visualize-your-assets.md
@@ -1,0 +1,5 @@
+# An asset library to visualize your assets
+
+Also included in Akeneo’s new Asset Manager features is an asset library that displays both assets stored in the PIM and those stored externally. This asset library offers full text search and filters.
+
+![The asset library screen](../img/Asset_families.png)
