+++ tests/updates/nominal_case/asset-manager/4-automatic-link-to-products.md
@@ -1,0 +1,5 @@
+# Automatically link assets to products with product link rules
+
+With the new Asset Manager, you can also create automatic links between your assets and your PIM products, whether assets are stored externally or internally.
+
+Product link rules help you save time in creating connections between assets and products. Each time a new asset is created in Akeneo, the rule is triggered and automatically links the asset to the right product!
