+++ tests/updates/nominal_case/asset-manager/5-asset-data-contextualization.md
@@ -1,0 +1,5 @@
+# Asset data contextualization
+
+Akeneo’s new Asset Manager also makes it possible to contextualize asset-related information!
+
+This new capability means you can have different attribute values (descriptions, names…), which vary by locale and/or channel. Asset values that differ by locale can be translated just like with other product attributes, helping you deliver an even better product experience.
