+++ tests/updates/nominal_case/asset-manager/6-transformation-rules.md
@@ -1,0 +1,6 @@
+# Asset transformation rules
+
+The new Asset Manager also offers additional flexibility in terms of asset transformations. You can, of course, keep your asset images as they are, distributing them in their original format without any transformation — but
+many times you need to transform the digital asset for different channel requirements. 
+
+The Asset Manager allows up to 10 transformations per asset family, and each transformation rendering will be visible from the asset form.
