+++ tests/updates/nominal_case/asset-manager/7-other-features.md
@@ -1,0 +1,16 @@
+# Other notable features
+
+## Permissions & completeness
+Assets are governed with permissions, and have dedicated completeness.
+
+![Asset completeness](../img/Asset_completeness.png)
+
+## Mass upload
+The asset mass upload screen has been revamped to offer a better user experience.
+
+![Asset completeness](../img/Assets_mass_upload.png)
+
+## New dedicated tab in the porduct form
+Assets are gathered under a new dedicated tab in the product form, where you will also be able to see all linked products for each asset.
+
+![Assets in the Product form](../img/Asset_in_PEF.png)
